consiolidates app sub models into app; adds app rules
