+++ pkg/models/app.go
@@ -2,6 +2,15 @@ package models
 
 import "time"
 
+type AppsQuery struct {
+	Limit       string
+	Page        string
+	Name        string
+	ConnectorID string
+	AuthMethod  string
+	Cursor      string
+}
+
 // App is the contract for apps api v2.
 type App struct {
 	ID                 *int32                   `json:"id,omitempty"`
@@ -24,11 +33,76 @@ type App struct {
 	RoleIDs            []int                    `json:"role_ids"`
 }
 
-type AppsQuery struct {
-	Limit       string
-	Page        string
-	Name        string
-	ConnectorID string
-	AuthMethod  string
-	Cursor      string
+// AppProvisioning is the contract for provisioning.
+type AppProvisioning struct {
+	Enabled *bool `json:"enabled,omitempty"`
+}
+
+// AppSso is the contract for apps sso.
+type AppSso struct {
+	ClientID     *string            `json:"client_id,omitempty"`
+	ClientSecret *string            `json:"client_secret,omitempty"`
+	MetadataURL  *string            `json:"metadata_url,omitempty"`
+	AcsURL       *string            `json:"acs_url,omitempty"`
+	SlsURL       *string            `json:"sls_url,omitempty"`
+	Issuer       *string            `json:"issuer,omitempty"`
+	Certificate  *AppSsoCertificate `json:"certificate"`
+}
+
+// AppSsoCertificate is the contract for sso certificate.
+type AppSsoCertificate struct {
+	ID    *int32  `json:"id,omitempty"`
+	Name  *string `json:"name,omitempty"`
+	Value *string `json:"value,omitempty"`
+}
+
+// AppConfiguration is the contract for configuration.
+type AppConfiguration struct {
+	RedirectURI                   *string `json:"redirect_uri,omitempty"`
+	RefreshTokenExpirationMinutes *int32  `json:"refresh_token_expiration_minutes,omitempty"`
+	LoginURL                      *string `json:"login_url,omitempty"`
+	OidcApplicationType           *int32  `json:"oidc_application_type,omitempty"`
+	TokenEndpointAuthMethod       *int32  `json:"token_endpoint_auth_method,omitempty"`
+	AccessTokenExpirationMinutes  *int32  `json:"access_token_expiration_minutes,omitempty"`
+	ProviderArn                   *string `json:"provider_arn,omitempty"`
+	SignatureAlgorithm            *string `json:"signature_algorithm,omitempty"`
+}
+
+// AppParameters is the contract for parameters.
+type AppParameters struct {
+	ID                        *int32  `json:"id,omitempty"`
+	Label                     *string `json:"label,omitempty"`
+	UserAttributeMappings     *string `json:"user_attribute_mappings,omitempty"`
+	UserAttributeMacros       *string `json:"user_attribute_macros,omitempty"`
+	AttributesTransformations *string `json:"attributes_transformations,omitempty"`
+	SkipIfBlank               *bool   `json:"skip_if_blank,omitempty"`
+	Values                    *string `json:"values,omitempty,omitempty"`
+	DefaultValues             *string `json:"default_values,omitempty"`
+	ParamKeyName              *string `json:"param_key_name,omitempty"`
+	ProvisionedEntitlements   *bool   `json:"provisioned_entitlements,omitempty"`
+	SafeEntitlementsEnabled   *bool   `json:"safe_entitlements_enabled,omitempty"`
+	IncludeInSamlAssertion    *bool   `json:"include_in_saml_assertion,omitempty"`
+}
+
+// AppRule is the contract for App Rules.
+type AppRule struct {
+	ID         *int32              `json:"id,omitempty"`
+	Name       *string             `json:"name,omitempty"`
+	Match      *string             `json:"match,omitempty"`
+	Enabled    *bool               `json:"enabled,omitempty"`
+	Position   *int32              `json:"position,omitempty"`
+	Conditions []AppRuleConditions `json:"conditions,omitempty"`
+	Actions    []AppRuleActions    `json:"actions,omitempty"`
+}
+
+type AppRuleConditions struct {
+	Source   *string `json:"source,omitempty"`
+	Operator *string `json:"operator,omitempty"`
+	Value    *string `json:"value,omitempty"`
+}
+
+type AppRuleActions struct {
+	Action     *string  `json:"action,omitempty"`
+	Value      []string `json:"value,omitempty"`
+	Expression *string  `json:"expression,omitempty"`
 }
