--- pkg/models/app_provisioning.go
@@ -1,6 +1,0 @@
-package models
-
-// AppProvisioning is the contract for provisioning.
-type AppProvisioning struct {
-	Enabled *bool `json:"enabled,omitempty"`
-}
