--- pkg/models/app_sso_certificate.go
@@ -1,8 +1,0 @@
-package models
-
-// AppSsoCertificate is the contract for sso certificate.
-type AppSsoCertificate struct {
-	ID    *int32  `json:"id,omitempty"`
-	Name  *string `json:"name,omitempty"`
-	Value *string `json:"value,omitempty"`
-}
